+++ clang/docs/ReleaseNotes.rst
@@ -135,6 +135,8 @@ C++ Language Changes
 C++2c Feature Support
 ^^^^^^^^^^^^^^^^^^^^^
 
+- Implemented `P3865R1 <https://wg21.link/P3865R1>`_ CTAD for type template template parameters.
+
 C++23 Feature Support
 ^^^^^^^^^^^^^^^^^^^^^